scraper/requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
alembic>=1.13.0
psycopg2-binary>=2.9.9
